Provide a portable way to wait for a file to be modified.
On most operating systems (Linux, illumos, macOS, *BSD, Windows), there
is a way to be woken up if another process writes to a file.  This can
be used to make a few places where we poll a file in a loop more
efficient.

The fallback behavior is simply to time out (Solaris, AIX).