Use pg_wait_file() in pg_ctl.
We wait for the postmaster to come up by reading the pid file repeatedly
with a sleep in between.  If possible, ask the OS to tell us when the
file has been written to, so we can wake up a bit sooner.

XXX This is wrong.  The file might be unlinked and replaced.  We need to
detect that by watching the parent directory.  And then we also need to
watch the newly created file.