ecpg: Use thread-safe _l() functions if possible.
In order to use snprintf("%g") and strtod() with the C locale in certain
parts of the ECPG code, we call uselocale() around those sections.

Two operating systems don't have uselocale(), but our fallbacks were
non-ideal.  For NetBSD we'd use setlocale(), which is not thread-safe
and dangerous, and for Windows we'd either use setlocale() or a Windows
API that is egregiously different and hard to maintain, depending on the
compiler tool chain.

We can remove all of that and use not-yet-standardized [v]s[n]printf_l()
and strtod_l(), if available.  Those two operating systems, along with
the other BSDs and macOS have them.  Glibc has strtod_l() but not yet
the other.

Discussion: https://postgr.es/m/CWZBBRR6YA8D.8EHMDRGLCKCD%40neon.tech