Vectored ring buffer writes.
When using eg BAS_VACUUM, we usually finish up writing recently dirtied
pages out to disk.  If the blocks are sequential, we can try to write
out multiple dirty buffers in ring order, to clean them in bulk.

XXX Should we try to detect cases when it's not working and suppress the
behaviour (ie if the dirty pages are in random order)?

XXX The reported statistics need work