Use streaming I/O in COPY via write-behind.
Feed dirty pinned buffers back to the buffer access strategy.  Not that
this only applies to simple heap-only COPY for now.  Index insertions
would require more effort, to make a strategy available to the right
code paths.  (In other words, the index insertion part of COPY has never
used BAS_WRITE and is not scan resistant, which we should fix.)