Use streaming I/O in BAS to do write-behind.
freelist.c traditionally peformed a sort of one-by-one write-behind, by
letting bufmgr.c clean the oldest buffer in the ring at the last moment
before recycling it.

Introduce a more explicit kind of write-behind, that begins once the
consumer of buffers has dirtied a buffer and is ready to unpin it.
Instead of calling ReleaseBuffer(), a new call StrategyReleaseBuffer()
can be used to feed dirty buffers back to the strategy as soon as
possible, so that it can feed them into a write stream.  The write
stream will release the pin some time later.

The main benefit today is that the write stream generates vectored I/O
up to io_combine_limit, and attempts to avoid stalling to flush the WAL
by re-ordering.  Using a write stream also insulates the code from
future work to support asynchronous I/O as required for direct I/O.

To actually benefit from this, alter commits will add
StrategyReleaseBuffer() to various places that dirty buffers in bulk.