Use latch for bgworker state change notification.
As part of a project to remove the reliance on PIDs to identify
backends, backends that register dynamic background workers will now
have their latch set by the postmaster when the worker state changes.
Previously, the postmaster would send a SIGUSR1 signal, which
relied on the ProcSignal system's handler setting the latch, with the
same end effect.  Now that intermediate step is skipped.

The field bgw_notify_pid still exists for backwards compatibility, but
has no effect and may be removed in a later release.
RegisterBackgroundWorker() now automatically assumes that the caller
would like state change notifications delivered to its proc number,
unless BGW_NO_NOTIFY is set in bgw_flags.

Removing other cases of PIDs from the bgwork API is left for later work;
this patch is concerned only with removing a dependency on ProcSignal
infrastructure that is due to be removed by a later commit.

This represents the first use of SetLatch() in the postmaster.  It might
seem more natural to use condition variables in the
wait-for-state-change functions, so that anyone with a handle can wait,
but condition variables as currently implemented would be a lot less
robust that simple latches.

Reviewed-by: Heikki Linnakangas <hlinnaka@iki.fi>
Discussion: https://postgr.es/m/CA%2BhUKG%2B3MkS21yK4jL4cgZywdnnGKiBg0jatoV6kzaniBmcqbQ%40mail.gmail.com