Use asynchronous socket I/O in backend.
Instead of waiting for readiness, run asynchronous I/O operations and
wait for completion.  This requires redesigning the buffering system, so
that we can be consuming from one buffer while the kernel is writing new
data into another, or the other way around.

The buffers used for socket data are controlled by new GUCs:

 * socket_buffers = the number of buffers
 * socket_buffer_size = the size of each buffer, default 8kB
 * socket_aio = use native socket AIO (if support available)

XXX WIP!