Introduce tiny local LRU cache for buffer mapping.
Remember which buffer held known recently accessed blocks for each fork
of each relation, so we can try to skip the shared memory buffer mapping
table, and more importantly its partition locks.

Access methods have to opt into this by classifying some reads as 'hot'
or 'warm'.  We have space for a memory of the last 8 blocks in each of
those categories, a size that was chosen because it fits in a cache line
for fast access.

ReadBufferExtended() takes new modes RBM_MAPPING_{HOT,WARM}, and for
StartReadBuffers() it's READ_BUFFERS_MAPPING_{HOT,WARM}.