jit: Add JITLink support.
Per https://llvm.org/docs/JITLink.html, JITLink is the preferred way to
link objects, and replaces RuntimeDyld.  Given our range of supported
LLVM versions, we'll need to support both for a time.

Since it has been announced that LLVM 20 will begin to deprecate the
older API, switch to JITLink for LLVM 20+.  Since the default
RuntimeDyld implementation has a known bug causing crashes on aarch64,
use JITLink for LLVM 14+ on that architecture.

The code change to select JITLink as the ObjectLinkingLayer is trivial,
but a small amount of new C++ code is currently needed in llvm_wrap.cpp
to provide access to JITLink from C code.  That could be proposed
upstream to LLVM to remove it from our tree eventually.  If doing so, it
would probably be slightly shorter and sweeter with C++14
std::make_unique (we restrict ourselves to C++11 for now).

XXX WIP