pgbench: Rationalize types in parseScriptWeight().
This function wants to parse a 64 bit number, but complain if it's out
of range for int, and include the number in the error message.  OK, but
long is not always bigger than int, so it won't work the same on all
systems (and the cast to long long in the error message won't really
make it longer).

Parse an int64 with strtoi64(), and print it out with PRId64, so now it
will behave the same everywhere.