Deprecate pg_int64 in libpq-fe.h.
Previously we used pg_int64 as a typename in the function prototypes of
libpq-fe.h.  It was set to the type that we had selected for int64 in
the backend.  That type is now always int64_t.

Given that it didn't match the name we use in the backend anyway, let's
just use int64_t directly in public interfaces.  It is likely to be the
name that modern clients application would use for their own 64-bit
processing anyway.

Keep a typedef marked deprecated for backward compatibility, but move it
to libpq-fe.h, where the only uses of it were.