Remove historical slashes from LSN format.
Ancient versions had a two-part XLogRecPtr where the parts had a
semantic meaning, and that was reflected in the display format.  Now
it's just a 64 bit number, so lets display it that way and remove most
remaining conversions to and from two-part form.

Unfortunately we can't defined a macro like LSN_FORMAT if we want
messages to be translatable.  Open-code formatting using <inttypes.h>
macros that the localization software understands.

XXX pg_upgrade probably needs to understand old format LSNs