XXX: Adaptive I/O depth control for read_stream.c.
XXX Experimental prototype science project code, not a proposal
XXX No claim that this is good enough for any purpose
XXX Works OK for easy cases and maybe it lights up the problem space

Whenever I/O is needed, we previously always boosted (doubled) the
look-ahead distance up to the allowed limit.  This rapidly achieved I/O
combining and concurrency and avoided I/O stalls, but it had no downward
pressure due to lack of information about the effectiveness of its
efforts.  The I/O queue would look something like this over time, with
one dot per queued I/O:

  .                                  }-- boost on I/O
  . .                                }
  . . . .                            }
  . . . . . . .                      }
  . . . . . . . . . . . . . . . .    }- settings prevent further growth
  . . . . . . . . . . . . . . . .    }
  . . . . . . . . . . . . . . . .    }

It reduces only on cache hits in PostgreSQL's own buffer pool.  If you
set the GUCs very high, you could use some extra resources (at least
pinned buffers and speculative work done in callback functions, also
I/Os and proposed asynchronous I/O workers processes during the initial
ramp-up phase) for no gain, though only when I/O is really needed.
Once you reach the maximum allowed size, the buffer usage plateaus and
the rate and concurrency of I/Os is bound by the rate of buffer
consumption from the stream, which is as it should be.

This experiment aims to see if we can scale it back without losing I/O
performance.  Some basic goals:

 * Trim distance only when we have evidence that we won't stall
 * Consider future potential I/O waits only, so we have time to react
 * Allow enough leeway to be wrong without stalls
 * Recover quickly when wrong and remember for a while
 * Avoid unending saw-tooth-like searches for the correct level
 * Adapt to changes in stream patterns rapidly
 * Work with the new asynchronous I/O APIs with no extra system calls
 * Have a plausible implementation even for synchronous I/O
 * Avoid complicated floating point logic

This commit adds a stall detection window consisting of two sample I/Os
at the 25% and 50% points in the current I/O queue, and checks whether
they would currently stall *if* we were to wait for them now (though we
won't yet).  If the near sample would, it boosts the distance as before.
If it wouldn't, but the far sample also wouldn't, then it begins to
adjust the distance downwards: I/Os are completing long before we need
them to, or the data was already well cached by the kernel, so we can
try to use fewer resources:

  .                      --- boost on minimum I/O startup
  .[S                    --- boost on stall warning
  .[n s].                }-- hold on window = no-stall, stall
  .[n s].                }
  .[n s].                }
  .[n s].                }
  .[S .].                --- boost on stall warning
  . .[n . N]. . .        }-- decay on excessive distance indicator
  . .[n . N]. . .        }
  . .[n . N]. . .        }
  . .[n . s]. .          }-- hold on window = no-stall, stall
  . .[n . s]. .          }
  . .[n . s]. .          }

Actual interesting queue depth may be considerably higher than that
sketch (XXX erm which is not quite right anyway, see later patch that
displays the real queue like this), but the window scales with it.
Whether that actually makes sense and where 25% and 50% came from would
be valid questions... well, it's easy and fast integer stuff...

This is in the general area of adaptive control theory which has
textbook algorithms and it is entirely possible that something with
industrial pedigree could produce much better results than the basic
window-triggered approach shown here.  This way does seem to match the
discrete "would stall" information we can cheaply get our hands on
pretty well, though it's possible that that could be converted into
something more like a continuous error variable if we sampled with with
running averages or something; but that's all slow floating point...

This patch only adds the adaptive algorithm, which is mostly copious
comments explaining the thought process and just a few lines of C code.
It doesn't yet add the feedback information required to feed it (see
next patches).  The dummy implementation of WaitReadBuffersMightStall()
here always returns true, meaning that it will always boost.  In other
words, as of this commit there is no change in behavior.